refactor: extract right inspector panel component
